// src/components/admin/types/analytics.ts
export interface Analytics {
  usersTotal: number;
  usersTrend: string;
  activeToday: number;
  activeTodayTrend: string;
  newLast7Days: number;
  newLast7DaysTrend: string;
  contentTotal: number;
  contentTrend: string;
  viewsToday: number;
  viewsTrend: string;
  commentsToday: number;
  commentsTrend: string;
}